Extract the select entry dropdown options under the file size limit.
